+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fetch target actor when processing portable `Follow` activity.
+- Add incoming activities to portable actor's inbox even if audience is not specified.
 
 ## [4.2.1] - 2025-05-08